+++ java-api/src/main/java/com/sap/cloud/security/config/ClientCertificate.java
@@ -6,12 +6,16 @@
 package com.sap.cloud.security.config;
 
 import javax.annotation.Nullable;
+import java.security.PrivateKey;
+import java.security.cert.Certificate;
 import java.util.Objects;
 
 public class ClientCertificate implements ClientIdentity {
 
-	private final String certificate;
-	private final String key;
+	private String certificate;
+	private Certificate[] certificateChain;
+	private PrivateKey privateKey;
+	private String key;
 	private final String clientId;
 
 	/**
@@ -28,6 +32,33 @@ public class ClientCertificate implements ClientIdentity {
 		this.certificate = certificate;
 		this.key = key;
 		this.clientId = clientId;
+	}
+
+	/**
+	 * Represents certificate based client identity.
+	 *
+	 * @param certificateChain
+	 * 		Certificate chain of the OAuth 2.0 client
+	 * @param privateKey
+	 * 		Private key of the OAuth 2.0 client
+	 * @param clientId
+	 * 		ID of the OAuth 2.0 client requesting the token.
+	 */
+	public ClientCertificate(@Nullable Certificate[] certificateChain, @Nullable PrivateKey privateKey,
+			@Nullable String clientId) {
+		this.certificateChain = certificateChain;
+		this.privateKey = privateKey;
+		this.clientId = clientId;
+	}
+
+	@Override
+	public Certificate[] getCertificateChain() {
+		return certificateChain;
+	}
+
+	@Override
+	public PrivateKey getPrivateKey() {
+		return privateKey;
 	}
 
 	@Override