+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## 3.4.4
+
+## 3.5.0
+- [java-api] `ClientIdentity` interface has been extended with 2 new methods `getCertificateChain()`
+  and `getPrivateKey()`
+  and `ClientCertificate` class has been extended with new constructor that takes `java.security.cert.Certificate[]`
+  and `java.security.PrivateKey` as an argument and corresponding getters for these fields.
+- [token-client] `SSLContextFactory` class has been extended and supports Keys in PKCS#8 format with ECC algorithm.
 - [spring-security] fixed NPE in IdentityServicesPropertySourceFactory on application startup when bound to a list of XSUAA services whose service plans are ALL not supported
+
 
 ## 3.4.3
 - [spring-security] improved custom SecurityContextStrategy registration for the `SecurityContextAutoConfiguration` class. It uses `ServletContextInitializer` to hook early into the initialization phase.